+++ ICD.Profound.ConnectPRO/Themes/UserInterface/Presenters/Common/Sources/SourceSelectPresenter.cs
@@ -13,7 +13,6 @@ using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IPresenters.Common.Sources;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews;
 using ICD.Profound.ConnectPRO.Themes.UserInterface.IViews.Common.Sources;
-using ICD.Profound.ConnectPRO.Themes.UserInterface.Views.Common.Sources;
 
 namespace ICD.Profound.ConnectPRO.Themes.UserInterface.Presenters.Common.Sources
 {